Move shared components into the /Components directory
