+++ frontend/src/Components/Charts/BarChart.tsx
